Refactor utility imports to consolidate module structure and add new dice rolling functionality
